+++ Assets/Scripts/GlockSystem.cs
@@ -6,7 +6,7 @@ public class GlockSystem : MonoBehaviour
     // i'll use this one to manage the stats of every weapon in game, this one is for glock
     WeaponController weaponControl;
     public float Glock_range = 100f;
-    public int Glock_Damage = 10; 
+    public float Glock_Damage = 10; 
     public float spread = 0.05f;
     public int maxBullets = 15;
     public string ReloadName = "Glock_Reload";
